Moved HTTP and gRPC to their own files
