+++ main.go
@@ -6,24 +6,15 @@ import (
 	"net"
 	"net/http"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/iver-wharf/wharf-core/pkg/cacertutil"
-	"github.com/iver-wharf/wharf-core/pkg/ginutil"
 	"github.com/iver-wharf/wharf-core/pkg/logger"
 	"github.com/iver-wharf/wharf-core/pkg/logger/consolepretty"
 	"github.com/soheilhy/cmux"
-	"google.golang.org/grpc"
+	"gorm.io/gorm"
 
-	"github.com/gin-contrib/cors"
-	"github.com/gin-gonic/gin"
-
-	v5 "github.com/iver-wharf/wharf-api/v5/api/wharfapi/v5"
 	"github.com/iver-wharf/wharf-api/v5/docs"
-	"github.com/iver-wharf/wharf-api/v5/internal/deprecated"
-	ginSwagger "github.com/swaggo/gin-swagger"
-	"github.com/swaggo/gin-swagger/swaggerFiles"
 )
 
 var log = logger.NewScoped("WHARF")
@@ -67,87 +58,11 @@ func main() {
 
 	seed()
 
-	db, err := openDatabase(config.DB)
-	if err != nil {
-		log.Error().
-			WithString("driver", string(config.DB.Driver)).
-			WithError(err).
-			Message("Database error")
-		os.Exit(2)
-	}
+	db := setupDB(config.DB)
+	serve(config, db)
+}
 
-	err = runDatabaseMigrations(db, config.DB.Driver)
-	if err != nil {
-		log.Error().WithError(err).Message("Migration error")
-		os.Exit(3)
-	}
-
-	gin.DefaultWriter = ginutil.DefaultLoggerWriter
-	gin.DefaultErrorWriter = ginutil.DefaultLoggerWriter
-
-	r := gin.New()
-	r.Use(
-		ginutil.LoggerWithConfig(ginutil.LoggerConfig{
-			//disable GIN logs for path "/health". Probes won't clog up logs now.
-			SkipPaths: []string{"/health"},
-		}),
-		ginutil.RecoverProblem,
-	)
-
-	if len(config.HTTP.CORS.AllowOrigins) > 0 {
-		log.Info().
-			WithStringf("origin", "%v", config.HTTP.CORS.AllowOrigins).
-			Message("Allowing origins in CORS.")
-		corsConfig := cors.DefaultConfig()
-		corsConfig.AllowOrigins = config.HTTP.CORS.AllowOrigins
-		corsConfig.AddAllowHeaders("Authorization")
-		corsConfig.AllowCredentials = true
-		r.Use(cors.New(corsConfig))
-	} else if config.HTTP.CORS.AllowAllOrigins {
-		log.Info().Message("Allowing all origins in CORS.")
-		corsConfig := cors.DefaultConfig()
-		corsConfig.AllowAllOrigins = true
-		r.Use(cors.New(corsConfig))
-	}
-
-	healthModule{}.DeprecatedRegister(r)
-	healthModule{}.Register(r.Group("/api"))
-
-	if config.HTTP.OIDC.Enable {
-		rsaKeys, err := GetOIDCPublicKeys(config.HTTP.OIDC.KeysURL)
-		if err != nil {
-			log.Error().WithError(err).Message("Failed to obtain OIDC public keys.")
-			os.Exit(1)
-		}
-		m := newOIDCMiddleware(rsaKeys, config.HTTP.OIDC)
-		r.Use(m.VerifyTokenMiddleware)
-		m.SubscribeToKeyURLUpdates()
-	}
-
-	setupBasicAuth(r, config)
-
-	modules := []httpModule{
-		engineModule{CIConfig: &config.CI},
-		branchModule{Database: db},
-		buildModule{Database: db, Config: &config},
-		projectModule{Database: db},
-		providerModule{Database: db},
-		tokenModule{Database: db},
-		deprecated.BranchModule{Database: db},
-		deprecated.BuildModule{Database: db},
-		deprecated.ProjectModule{Database: db},
-		deprecated.ProviderModule{Database: db},
-		deprecated.TokenModule{Database: db},
-	}
-
-	api := r.Group("/api")
-	for _, module := range modules {
-		module.Register(api)
-	}
-
-	api.GET("/version", getVersionHandler)
-	api.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
-
+func serve(config Config, db *gorm.DB) {
 	listener, err := net.Listen("tcp", config.HTTP.BindAddress)
 	if err != nil {
 		log.Error().WithError(err).
@@ -159,12 +74,8 @@ func main() {
 		cmux.HTTP2MatchHeaderFieldSendSettings("content-type", "application/grpc"))
 	httpListener := mux.Match(cmux.Any())
 
-	grpcServer := grpc.NewServer()
-	grpcWharf := &grpcWharfServer{db: db}
-	v5.RegisterBuildsServer(grpcServer, grpcWharf)
-
-	go grpcServer.Serve(grpcListener)
-	go r.RunListener(httpListener)
+	go serveGRPC(grpcListener, db)
+	go serveHTTP(httpListener, config, db)
 
 	if err := mux.Serve(); err != nil {
 		log.Error().WithError(err).
@@ -172,27 +83,23 @@ func main() {
 	}
 }
 
-func setupBasicAuth(router *gin.Engine, config Config) {
-	if config.HTTP.BasicAuth == "" {
-		log.Info().Message("BasicAuth setting not set, skipping BasicAuth setup.")
-		return
+func setupDB(dbConfig DBConfig) *gorm.DB {
+	db, err := openDatabase(dbConfig)
+	if err != nil {
+		log.Error().
+			WithString("driver", string(dbConfig.Driver)).
+			WithError(err).
+			Message("Database error")
+		os.Exit(2)
 	}
 
-	accounts := gin.Accounts{}
-	var accountNames []string
-
-	for _, account := range strings.Split(config.HTTP.BasicAuth, ",") {
-		split := strings.Split(account, ":")
-		user, pass := split[0], split[1]
-
-		accounts[user] = pass
-		accountNames = append(accountNames, user)
+	err = runDatabaseMigrations(db, dbConfig.Driver)
+	if err != nil {
+		log.Error().WithError(err).Message("Migration error")
+		os.Exit(3)
 	}
 
-	log.Debug().WithString("usernames", strings.Join(accountNames, ",")).
-		Messagef("Set up basic authentication for %d users.", len(accountNames))
-
-	router.Use(gin.BasicAuth(accounts))
+	return db
 }
 
 func seed() {
